fixing some bugs
changing the code in a way that the fruit drop position doesn't collide with the current snake, adding a head and tail, and changing some visual features such as background color and snake color. snake will lose right after it collides with the wall not when it is outside the game surface.